make possible to disable rate limit

+++ cmd/access/node_builder/access_node_builder.go
@@ -1480,7 +1480,7 @@ func (builder *FlowAccessNodeBuilder) extraFlags() {
 			defaultConfig.rpcConf.WebSocketConfig.MaxSubscriptionsPerConnection,
 			"the maximum number of active WebSocket subscriptions allowed per connection",
 		)
-		flags.Uint64Var(
+		flags.Float64Var(
 			&builder.rpcConf.WebSocketConfig.MaxResponsesPerSecond,
 			"websocket-max-responses-per-second",
 			defaultConfig.rpcConf.WebSocketConfig.MaxResponsesPerSecond,
